# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reference/categories": {
      "get": {
        "operationId": "ReferenceController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reference"
        ]
      },
      "post": {
        "operationId": "ReferenceController_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "reference"
        ]
      }
    },
    "/reference/units": {
      "get": {
        "operationId": "ReferenceController_getUnits",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UnitResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "reference"
        ]
      },
      "post": {
        "operationId": "ReferenceController_createUnit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUnitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnitResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "reference"
        ]
      }
    },
...
    "/profile/check-username": {
      "get": {
        "operationId": "ProfileController_checkUsername",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckUsernameResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/me": {
      "get": {
        "operationId": "ProfileController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ]
      },
      "patch": {
        "operationId": "ProfileController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/me/settings": {
      "get": {
        "operationId": "ProfileController_getMySettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSettingsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ]
      },
      "patch": {
        "operationId": "ProfileController_updateMySettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileSettingsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/me/followers": {
      "get": {
        "operationId": "ProfileController_getMyFollowers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProfilesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/me/following": {
      "get": {
        "operationId": "ProfileController_getMyFollowing",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProfilesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ]
      }
    },
    "/profile/me/following/{userId}": {
      "post": {
        "operationId": "ProfileController_follow",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HabitProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ]
      },
      "delete": {
        "operationId": "ProfileController_unfollow",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "Habit Tracker API",
    "description": "API for habit tracking, community, challenges and gamification",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iconName": {
            "type": "object"
          },
          "colorValue": {
            "type": "object"
          },
          "imageUrl": {
            "type": "object"
          },
          "isPredefined": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "isPredefined"
        ]
      },
      "UnitResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "isPredefined": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "symbol",
          "isPredefined"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Health"
          },
          "iconName": {
            "type": "object"
          },
          "colorValue": {
            "type": "object"
          },
          "imageUrl": {
            "type": "object"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateUnitDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Kilometers"
          },
          "symbol": {
            "type": "string",
            "example": "km"
          }
        },
        "required": [
          "name",
          "symbol"
        ]
      },
      "CheckUsernameResponseDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean"
          }
        },
        "required": [
          "available"
        ]
      },
      "HabitProfileResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "subscriptionTier": {
            "type": "string",
            "enum": [
              "FREE",
              "PRO",
              "LIFETIME"
            ]
          },
          "bio": {
            "type": "object"
          },
          "points": {
            "type": "number"
          },
          "isTaggingAllowed": {
            "type": "boolean"
          },
          "avatarUrl": {
            "type": "object"
          },
          "followerCount": {
            "type": "number"
          },
          "followingCount": {
            "type": "number"
          }
        },
        "required": [
          "userId",
          "username",
          "subscriptionTier",
          "points",
          "isTaggingAllowed"
        ]
      },
      "GetMeResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object"
          },
          "profile": {
            "$ref": "#/components/schemas/HabitProfileResponseDto"
          }
        },
        "required": [
          "user",
          "profile"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "bio": {
            "type": "string",
            "maxLength": 500
          },
          "isTaggingAllowed": {
            "type": "boolean"
          },
          "avatarUrl": {
            "type": "object"
          }
        }
      },
      "ProfileSettingsResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "isSearchable": {
            "type": "boolean"
          },
          "analyticsEnabled": {
            "type": "boolean"
          },
          "profileVisibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "challengeVisibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "challengePostVisibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "taskDailyReminderTime": {
            "type": "object"
          },
          "taskWeekStartDay": {
            "type": "number"
          },
          "taskArchiveVisible": {
            "type": "boolean"
          },
          "pomodoroFocusDuration": {
            "type": "number"
          },
          "pomodoroBreakDuration": {
            "type": "number"
          },
          "pomodoroLongBreakDuration": {
            "type": "number"
          }
        },
        "required": [
          "userId",
          "isSearchable",
          "analyticsEnabled",
          "profileVisibility",
          "challengeVisibility",
          "challengePostVisibility",
          "taskWeekStartDay",
          "taskArchiveVisible",
          "pomodoroFocusDuration",
          "pomodoroBreakDuration",
          "pomodoroLongBreakDuration"
        ]
      },
      "UpdateProfileSettingsDto": {
        "type": "object",
        "properties": {
          "isSearchable": {
            "type": "boolean"
          },
          "analyticsEnabled": {
            "type": "boolean"
          },
          "profileVisibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "challengeVisibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "challengePostVisibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "taskDailyReminderTime": {
            "type": "object"
          },
          "taskWeekStartDay": {
            "type": "number",
            "minimum": 0,
            "maximum": 6
          },
          "taskArchiveVisible": {
            "type": "boolean"
          },
          "pomodoroFocusDuration": {
            "type": "number",
            "minimum": 1
          },
          "pomodoroBreakDuration": {
            "type": "number",
            "minimum": 1
          },
          "pomodoroLongBreakDuration": {
            "type": "number",
            "minimum": 1
          }
        }
      },
      "PaginatedProfilesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HabitProfileResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}